refactor(frontend): extract createPriceChartSeries factory from chart init
Move chart + all 6 series construction out of CandlestickChart's init
useLayoutEffect into a focused createPriceChartSeries factory module.
No behavior change — same series order, options, colors, and scaleMargins.
